Move `VerifyState` to a separate package
